test: extract test suite from aasmanagermultithreading test
